requirement.txt
appdirs==1.4.4
APScheduler==3.6.3
astroid==2.4.2
attrs==20.3.0
Authlib==0.15.2
autopep8==1.5.4
awscli==1.18.196
backports.shutil-get-terminal-size==1.0.0
backports.shutil-which==3.5.2
beautifulsoup4==4.9.3
botocore==1.19.36
braceexpand==0.1.6
cachelib==0.1.1
cairocffi==1.2.0
CairoSVG==2.5.0
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
check50==3.2.0
click==7.1.2
colorama==0.4.3
compare50==1.2.2
cryptography==3.3.1
cs50==6.0.2
cssselect2==0.4.1
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.1
docutils==0.15.2
EditorConfig==0.12.2
filelock==3.0.12
Flask==1.1.2
Flask-Session==0.3.2
help50==3.0.3
html5lib==1.1
icdiff==1.9.1
idna==2.10
ikp3db==1.4.3
importlib-metadata==3.3.0
intervaltree==2.1.0
isort==5.6.4
itsdangerous==1.1.0
jellyfish==0.8.2
Jinja2==2.11.2
jmespath==0.10.0
joblib==1.0.0
jsbeautifier==1.13.0
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
lib50==2.0.9
MarkupSafe==1.1.1
matplotlib==3.3.3
mccabe==0.6.1
natsort==7.1.0
nltk==3.5
numpy==1.19.4
oauthlib==3.1.0
pandas==1.1.5
pexpect==4.8.0
Pillow==8.0.1
plotly==4.14.1
ptyprocess==0.6.0
pyasn1==0.4.8
pycodestyle==2.6.0
pycparser==2.20
Pygments==2.7.3
pylint==2.6.0
pylint-django==2.3.0
pylint-flask==0.6
pylint-plugin-utils==0.6
pyparsing==2.4.7
PyPDF2==1.26.0
Pyphen==0.10.0
pyTelegramBotAPI==3.7.4
python-dateutil==2.8.1
python-magic==0.4.18
python-telegram-bot==13.1
pytz==2020.4
PyYAML==5.3.1
regex==2020.11.13
render50==4.2.1
requests==2.25.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.5
s3cmd==2.1.0
s3transfer==0.3.3
six==1.15.0
sortedcontainers==2.3.0
soupsieve==2.1
SQLAlchemy==1.3.20
sqlparse==0.4.1
style50==2.7.4
submit50==3.0.2
termcolor==1.1.0
tinycss2==1.1.0
toml==0.10.2
tornado==6.1
tqdm==4.54.1
twython==3.8.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.2
virtualenv==20.2.2
WeasyPrint==52.2
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.0
